Added testing and refactored
