+++ Makefile
@@ -87,7 +87,7 @@ $U/initcode: $U/initcode.S
 tags: $(OBJS) _init
 	etags *.S *.c
 
-ULIB = $U/ulib.o $U/usys.o $U/printf.o $U/umalloc.o $U/vec.o $U/shuffler.o
+ULIB = $U/ulib.o $U/usys.o $U/printf.o $U/umalloc.o $U/vec.o $U/shuffler.o $U/shuf_start.o
 
 _%: %.o $(ULIB)
 	$(LD) $(LDFLAGS) -T $U/user.ld -o $@ $^
@@ -129,6 +129,7 @@ UPROGS=\
 	$U/_sh\
 	$U/_shuf\
 	$U/_stressfs\
+	$U/_test_shuf\
 	$U/_time\
 	$U/_usertests\
 	$U/_grind\
